+++ RELEASE_NOTES.md
@@ -54,3 +54,4 @@
 - Make the `nox-cross-arch-all` job fail if any `nox-cross-arch` matrix job fails.
 - Fix credentials not being passed to the `test-installation` job in the CI workflow.
 - Make sure credentials are configured for all jobs that check out the repository in the CI workflow.
+- Disable the new `check-class-attributes` check in pydoclint 0.5.3, as we use a different way to document class attributes.